payment smooth

+++ Ecommerce_sam/checkout_app/views.py
@@ -47,6 +47,7 @@ def payment(request):
         if request.user.is_authenticated:
             user = request.user
         else:
+
             if SiteUser.objects.filter(email=request.POST.get('billing_email')).count()>0:
                 request.POST = request.POST
                 if 'available_cart_pk' in request.session:
@@ -154,17 +155,28 @@ def payment(request):
         if request.user.is_authenticated:
             cart_latest = Cart_model.objects.filter(user_id=request.user.id, is_payment_done=False)
             max_count = cart_latest.count()
-            for i in range(1,max_count+1):
-                Cart_model.objects.filter(id=Cart_model.objects.filter(user_id=request.user.id, is_payment_done=False).order_by('-id')[i].id).delete()
-            cart_latest = Cart_model.objects.filter(user_id=request.user.id, is_payment_done=False).order_by('-id')[0]
+            if max_count >1:
+                for i in range(1,max_count+1):
+                    Cart_model.objects.filter(id=Cart_model.objects.filter(user_id=request.user.id, is_payment_done=False).order_by('-id')[i].id).delete()
+                cart_latest = Cart_model.objects.filter(user_id=request.user.id, is_payment_done=False).order_by('-id')[0]
 
-            cart_objs = Cart_products.objects.filter(cart_id=cart_latest.id)
+                cart_objs = Cart_products.objects.filter(cart_id=cart_latest.id)
+            elif max_count == 1:
+                cart_latest = Cart_model.objects.filter(user_id=request.user.id, is_payment_done=False).order_by('-id')[
+                    0]
+                cart_objs = Cart_products.objects.filter(cart_id=cart_latest.id)
+            else:
+                cart_objs = Cart_model.objects.filter(id=request.session['available_cart_pk'], is_payment_done=False).order_by('-id')[0]
+                Cart_model.objects.filter(id=cart_objs.id).update(user_id=request.user.id)
+                cart_objs = Cart_products.objects.filter(cart_id=cart_objs.id)
+
+
             cart_item_exist = True if cart_objs.count() > 0 else False
             total_product_cost = cart_objs.aggregate(Sum('product_cost'))
 
         elif 'available_cart_pk' in request.session:
             cart_objs = Cart_model.objects.filter(id=request.session['available_cart_pk'], is_payment_done=False).order_by('-id')[0]
-            cart_objs.update(user_id=request.user.id)
+            Cart_model.objects.filter(id=cart_objs.id).update(user_id=request.user.id)
             cart_objs = Cart_products.objects.filter(cart_id=cart_objs.id)
 
             cart_item_exist = True if cart_objs.count()>0 else False
